+++ src/core/attrs.js
@@ -5,7 +5,14 @@ const attrsDefault = {
     class: 'easy-class',
     rcoe: 'easy-rcoe',
     parallel: 'easy-parallel',
-    self: 'easy-self'
-}
+    self: 'easy-self',
+    selfRcoe: 'easy-self-rcoe',
+
+    // Hooks
+    onToggle($el){},
+    onAdd($el){},
+    onRemove($el){},
+    onRcoe($el){},
+};
 
 export default attrsDefault;