+++ src/core/index.js
@@ -1,6 +1,7 @@
 import toggle from './toggle';
 import add from './add';
 import remove from './remove';
+import rcoes from './rcoes';
 
 function easyTogglerHandler(e, attrs) {
     const $toggler = e.target.closest(`[${ attrs.toggle }]`);
@@ -23,21 +24,7 @@ function easyTogglerHandler(e, attrs) {
     }
 
     if(!$toggler && !$remove && !$add) {
-        const $rcoes = document.querySelectorAll(`[${ attrs.rcoe }]`);
-
-        Array.from($rcoes).forEach(($rcoe) => {
-            let block = $rcoe.getAttribute(attrs.toggle);
-            let block_class = $rcoe.getAttribute(attrs.class);
-
-            if(!e.target.closest(block)) {
-                document.querySelector(block)?.classList.remove(block_class);
-
-                const _selfClass = $rcoe.getAttribute(attrs.self);
-                if(_selfClass) {
-                    $rcoe.classList.remove(_selfClass);
-                }
-            }
-        });
+        rcoes(e, attrs);
     }
 }