+++ src/easy-toggler.umd.js
@@ -2,5 +2,7 @@ import easyTogglerHandler from './core';
 import attrsDefault from './core/attrs';
 
 document.addEventListener('DOMContentLoaded', () => {
-    document.addEventListener('click', (e) => easyTogglerHandler(e, attrsDefault));
+    document.addEventListener('click', (e) => {
+        easyTogglerHandler(e, attrsDefault);
+    });
 });